Modify project structure
