+++ pyteal/__init__.py
@@ -1,2 +1,4 @@
-from .ops import *
+from .ast import *
+from .errors import TealInternalError, TealTypeError, TealTypeMismatchError, TealInputError
 from .util import execute
+from .config import MAX_GROUP_SIZE
