+++ pyteal/ast/addr.py
@@ -1,0 +1,29 @@
+from typing import Union
+
+from ..types import TealType, valid_address
+from .leafexpr import LeafExpr
+from .tmpl import Tmpl
+
+class Addr(LeafExpr):
+    """An expression that represents an Algorand address."""
+     
+    def __init__(self, address: Union[str, Tmpl]) -> None:
+        """Create a new Addr expression.
+        
+        Args:
+            address: A string containing a valid base32 Algorand address, or a Tmpl object.
+        """
+        if isinstance(address, Tmpl):
+            self.address = address.name
+        else:
+            valid_address(address)
+            self.address = address
+
+    def __teal__(self):
+        return [["addr", self.address]]
+
+    def __str__(self):
+        return "(address: {})".format(self.address)
+
+    def type_of(self):
+        return TealType.bytes
