+++ pyteal/ast/leafexpr.py
@@ -1,0 +1,5 @@
+from .expr import Expr
+
+class LeafExpr(Expr):
+    """Leaf expression base class."""
+    pass
